+++ SampleTest.Tests/Tests.cs
@@ -26,6 +26,13 @@ namespace SampleTest.Tests
             Assert.Ignore("Ignore Message");
         }
 
+        [Test]
+        [Explicit]
+        public void TestExplicit()
+        {
+
+        }
+
         [Test(ExpectedResult = 1)]
         public int TestReturn()
         {
@@ -49,33 +56,20 @@ namespace SampleTest.Tests
             Assert.True(true);
         }
 
-        public static int MyProperty { get; set; }
-
-        [Test]
-        public void Test1()
-        {
-            Console.WriteLine($"Test1: {MyProperty++}");
-        }
-
-        [Test]
-        public void Test3()
-        {
-            Console.WriteLine($"Test3: {MyProperty++}");
-        }
-
-        [Test]
-        public void Test2()
-        {
-            Console.WriteLine($"Test2: {MyProperty++}");
-        }
-
-        [Test]
-        [TestCase(1, TestName = "TestName")]
-        [TestCase(2)]
+        [TestCase(1, TestName = "TestName1")]
+        [TestCase(2, TestName = "TestName2")]
         [TestCase(3)]
+        [TestCase(4)]
+        [TestCase(5, TestName = "TestSame")]
         public void TestCases(int i)
         {
             Assert.True(i > 0);
+        }
+
+        [Test]
+        public void TestSame()
+        {
+
         }
     }
 }