+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] / 2022-12-01
+### Features
+- Featere `TestEngine.GetTestFullNames` supported
+- Feature `WildcardPattern` filter supported
+- Feature `TestCaseAttribute` supported
+### Added
+- Add `Alias` in `TestModel`
+- Add `ExplicitEnabled` in Filter
+- Add `WildcardPattern`
+- Add `HasFilterTestMethod`
+- Update `GetFilterTestMethods`
+### Fixed
+- Fix `SetUp` & `TearDown`
+### Tests
+- Test Static with `SetUp` & `TearDown`
+
 ## [1.0.6] / 2022-11-30 - 2022-10-31
 ### Features
 - Add `TestEngineFilter` simple filter by method/test name
@@ -59,6 +75,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test with Console Output
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.7]: ../../compare/1.0.6...1.0.7
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5
 [1.0.4]: ../../compare/1.0.3...1.0.4